+++ TinyFFR/World/ISceneBuilder.cs
@@ -6,8 +6,14 @@ using System;
 namespace Egodystonic.TinyFFR.World;
 
 public interface ISceneBuilder {
-	Scene CreateScene(bool includeBackdrop = true, ColorVect? backdropColor = null, ReadOnlySpan<char> name = default) {
-		return CreateScene(new SceneCreationConfig { InitialBackdropColor = includeBackdrop ? (backdropColor ?? SceneCreationConfig.DefaultInitialBackdropColor) : null, Name = name});
+	Scene CreateScene(ReadOnlySpan<char> name = default) {
+		return CreateScene(new SceneCreationConfig { Name = name});
+	}
+	Scene CreateScene(ColorVect? backdropColor, ReadOnlySpan<char> name = default) {
+		return CreateScene(new SceneCreationConfig { InitialBackdropColor = backdropColor, Name = name});
+	}
+	Scene CreateScene(BuiltInSceneBackdrop backdrop, ReadOnlySpan<char> name = default) {
+		return CreateScene(new SceneCreationConfig { InitialBackdrop = backdrop, Name = name});
 	}
 	Scene CreateScene(in SceneCreationConfig config);
 }