+++ Documentation/source/changelog.md
@@ -21,6 +21,7 @@ __Github: [Issues](https://github.com/Egodystonic/TinyFFR/milestone/9?closed=1) 
 * Added support for orthographic camera projection.
 * Greatly improved Linux stability. First-class support for Wayland (X11 support is removed for now).
 * Improved loading speed for skybox textures on Linux
+* Added built-in scene backdrops
 
 ### Bug Fixes